+++ inc/contractcost.class.php
@@ -179,10 +179,30 @@ class ContractCost extends CommonDBChild {
       return $tab;
    }
 
+   /**
+    * get the request results to get items associated to the given one (defined by $itemtype and $items_id)
+    *
+    * @param string  $itemtype          the type of the item we want the resulting items to be associated to
+    * @param string  $items_id          the name of the item we want the resulting items to be associated to
+    *
+    * @return array the items associated to the given one (empty if none was found)
+   **/
+   static function getItemsAssociationRequest($itemtype, $items_id) {
+      global $DB;
+
+      return $DB->request([
+         'SELECT' => 'id',
+         'FROM'   => static::getTable(),
+         'WHERE'  => [
+         static::$items_id  => $items_id
+         ]
+      ]);
+   }
 
    /**
     * Duplicate all costs from a contract template to its clone
     *
+    * @deprecated 9.5
     * @since 0.84
     *
     * @param $oldid
@@ -191,6 +211,7 @@ class ContractCost extends CommonDBChild {
    static function cloneContract($oldid, $newid) {
       global $DB;
 
+      Toolbox::deprecated('Use clone');
       $iterator = $DB->request([
          'FROM'   => self::getTable(),
          'WHERE'  => ['contracts_id' => $oldid]