+++ inc/item_softwarelicense.class.php
@@ -892,20 +892,23 @@ JAVASCRIPT;
    /**
     * Duplicate all software licenses from a computer template to its clone
     *
+    * @deprecated 9.5
+    *
     * @param integer $oldid ID of the computer to clone
     * @param integer $newid ID of the computer cloned
     *
     * @return void
-    * @deprecated 9.5.0
    **/
    static function cloneComputer($oldid, $newid) {
-      Toolbox::deprecated('Use cloneItem');
+      Toolbox::deprecated('Use clone');
       self::cloneItem('Computer', $oldid, $newid);
    }
 
 
    /**
     * Duplicate all software licenses from an item template to its clone
+    *
+    * @deprecated 9.5
     *
     * @param string  $itemtype Type of the item
     * @param integer $oldid ID of the item to clone
@@ -916,6 +919,7 @@ JAVASCRIPT;
    static function cloneItem($itemtype, $oldid, $newid) {
       global $DB;
 
+      Toolbox::deprecated('Use clone');
       $iterator = $DB->request([
          'FROM' => 'glpi_items_softwarelicenses',
          'WHERE' => [