+++ tests/functionnal/MassiveAction.php
@@ -43,18 +43,18 @@ class MassiveAction extends DbTestCase {
          [
             'itemtype'     => 'Computer',
             'items_id'     => '_test_pc01',
-            'allcount'     => 14,
-            'singlecount'  => 7
+            'allcount'     => 15,
+            'singlecount'  => 8
          ], [
             'itemtype'     => 'Printer',
             'items_id'     => '_test_printer_all',
-            'allcount'     => 12,
-            'singlecount'  => 6
+            'allcount'     => 13,
+            'singlecount'  => 7
          ], [
             'itemtype'     => 'Ticket',
             'items_id'     => '_ticket01',
-            'allcount'     => 15,
-            'singlecount'  => 10
+            'allcount'     => 16,
+            'singlecount'  => 11
          ], [
             'itemtype'     => 'Profile',
             'items_id'     => 'Super-Admin',